fix(crawl): address adaptive policy review
